Move bash helper functions into seperate file
