+++ bash/gast_to_code/bash_gast_to_code_converter.py
@@ -1,6 +1,7 @@
 from shared.gast_to_code.abstract_gast_to_code_converter import AbstractGastToCodeConverter
 import shared.gast_to_code.general_helpers as general_helpers
 import shared.gast_to_code.gast_to_code_router as router
+import bash.gast_to_code.bash_helpers as bash_helpers
 
 
 class BashGastToCodeConverter(AbstractGastToCodeConverter):
@@ -21,7 +22,8 @@ class BashGastToCodeConverter(AbstractGastToCodeConverter):
                     gast["args"]):
             return "impossibleTranslationError: direct translation does not exist"  # TODO: streamline error messag
         return router.gast_to_code(gast["value"],
-                                   "bash") + " " + bash_arg_helper(gast["args"])
+                                   "bash") + " " + bash_helpers.bash_arg_helper(
+                                       gast["args"])
 
     def handle_arr(gast):
         # This logic returns an error for nested arrays which are not supported in bash
@@ -67,26 +69,3 @@ class BashGastToCodeConverter(AbstractGastToCodeConverter):
 
     def handle_root(gast):
         return general_helpers.list_helper(gast["body"], "bash", "\n")
-
-
-'''
-Helper functions for bash converter since variables are written different when function params
-These functions are called on func args only and return the correct translation
-'''
-
-
-def handle_var_arg(gast):
-    return '"$' + gast["value"] + '"'
-
-
-def bash_arg_helper(gast_list):
-    out = ""
-    for i in range(0, len(gast_list)):
-        if gast_list[i]["type"] == "name":
-            out += handle_var_arg(gast_list[i])
-        else:
-            out += router.gast_to_code(gast_list[i], "bash")
-
-        if i < len(gast_list) - 1:  # don't add delimiter for last item
-            out += " "
-    return out
